attempt 69 for T2
